refactored code, updated povray scripts
