+++ main.py
@@ -1,16 +1,20 @@
 import sys
-
 import cv2
-import exifread
-from GPSPhoto import gpsphoto
+from src.data_getters.mountains import get_mountain_data
 
 from src.debug_tools import p_a, p_i
 from src.feature_matching import feature_matching
 from src.location_handler import look_at_location
 from src.photo_filtering import photo_filtering, resizer
-from src.pov_generator import file_handler
+from src.dem import render_dem
+
 
 if __name__ == '__main__':
+    render_dem(*get_mountain_data('data/dem-data.json'))
+    
+
+
+    exit()
     try:
         mode = int(sys.argv[1])
         index = 2
@@ -21,7 +25,7 @@ if __name__ == '__main__':
     if mode == 0:
         p_i("Running the program in DEM mode")
         blåmanen = 60.4014, 5.3633
-        damsgårdsfjellet = 60.3774, 5.2914
+         
         fløyen = 60.3944, 5.3432
         gulfjellet = 60.3731, 5.5804
         løvstakken = 60.3609, 5.3191
@@ -37,6 +41,7 @@ if __name__ == '__main__':
         camera_lat, camera_lon = strandafjellet
         look_at_lat, look_at_lon = ulriken
 
+        """ 
         try:
             data = gpsphoto.getGPSData(sys.argv[index + 1])
             p_i("Image Lat/Lon: %f, %f" % (data['Latitude'], data['Longitude']))
@@ -50,7 +55,8 @@ if __name__ == '__main__':
                 heading = 0
             look_at_lat, look_at_lon = look_at_location(camera_lat, camera_lon, 5, heading)
         except IndexError:
-            pass
+            pass 
+        """
 
         try:
             file_handler(sys.argv[index], camera_lat, camera_lon, look_at_lat, look_at_lon)
